langchain/requirements.txt
langchain
langchain_openai
langchain-ollama
transformers
torch
flask  
requests
python-dotenv
